# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OpenKeep API",
    "description": "Open-source, self-hostable Google Keep alternative",
    "version": "1.28.0"
  },
...
  "paths": {
...
    "/api/notes/{id}/state": {
      "patch": {
        "tags": [
          "notes"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "pinned": {
                    "type": "boolean"
                  },
                  "archived": {
                    "type": "boolean"
                  },
                  "isTemplate": {
                    "type": "boolean"
                  },
                  "color": {
                    "type": "string",
                    "enum": [
                      "default",
                      "coral",
                      "peach",
                      "sand",
                      "mint",
                      "sage",
                      "fog",
                      "storm",
                      "dusk",
                      "blossom",
                      "clay",
                      "chalk"
                    ]
                  },
                  "background": {
                    "type": "string",
                    "enum": [
                      "none",
                      "groceries",
                      "food",
                      "music",
                      "recipes",
                      "notes",
                      "places",
                      "travel",
                      "video",
                      "celebration"
                    ]
                  },
                  "position": {
                    "type": "string",
                    "maxLength": 256
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "pinned": {
                      "type": "boolean"
                    },
                    "archived": {
                      "type": "boolean"
                    },
                    "isTemplate": {
                      "type": "boolean"
                    },
                    "color": {
                      "type": "string",
                      "enum": [
                        "default",
                        "coral",
                        "peach",
                        "sand",
                        "mint",
                        "sage",
                        "fog",
                        "storm",
                        "dusk",
                        "blossom",
                        "clay",
                        "chalk"
                      ]
                    },
                    "background": {
                      "type": "string",
                      "enum": [
                        "none",
                        "groceries",
                        "food",
                        "music",
                        "recipes",
                        "notes",
                        "places",
                        "travel",
                        "video",
                        "celebration"
                      ]
                    },
                    "position": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "pinned",
                    "archived",
                    "isTemplate",
                    "color",
                    "background",
                    "position"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/notes/{id}/trash": {
      "post": {
        "tags": [
          "notes"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "type": {
                      "type": "string",
                      "enum": [
                        "text",
                        "list"
                      ]
                    },
                    "title": {
                      "type": "string"
                    },
                    "bodyHtml": {
                      "type": "string"
                    },
                    "hasLinks": {
                      "type": "boolean"
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                          },
                          "text": {
                            "type": "string",
                            "maxLength": 1000
                          },
                          "checked": {
                            "type": "boolean"
                          },
                          "indent": {
                            "anyOf": [
                              {
                                "type": "number",
                                "enum": [
                                  0
                                ]
                              },
                              {
                                "type": "number",
                                "enum": [
                                  1
                                ]
                              }
                            ]
                          },
                          "position": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "text",
                          "checked",
                          "indent",
                          "position"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "labelIds": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                      }
                    },
                    "attachments": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                          },
                          "kind": {
                            "type": "string",
                            "enum": [
                              "image",
                              "audio",
                              "drawing",
                              "file"
                            ]
                          },
                          "mime": {
                            "type": "string"
                          },
                          "width": {
                            "nullable": true,
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          },
                          "height": {
                            "nullable": true,
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          },
                          "filename": {
                            "nullable": true,
                            "type": "string"
                          },
                          "hasThumb": {
                            "type": "boolean"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                          }
                        },
                        "required": [
                          "id",
                          "kind",
                          "mime",
                          "width",
                          "height",
                          "filename",
                          "hasThumb",
                          "createdAt",
                          "updatedAt"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "reminder": {
                      "nullable": true,
                      "type": "object",
                      "properties": {
                        "remindAt": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                        },
                        "rrule": {
                          "nullable": true,
                          "type": "string"
                        },
                        "timezone": {
                          "type": "string"
                        },
                        "snoozedUntil": {
                          "nullable": true,
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                        },
                        "done": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "remindAt",
                        "rrule",
                        "timezone",
                        "snoozedUntil",
                        "done"
                      ],
                      "additionalProperties": false
                    },
                    "collaborators": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "userId": {
                            "type": "string"
                          },
                          "email": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "role": {
                            "type": "string",
                            "enum": [
                              "viewer",
                              "collaborator",
                              "owner"
                            ]
                          }
                        },
                        "required": [
                          "userId",
                          "email",
                          "name",
                          "role"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "role": {
                      "type": "string",
                      "enum": [
                        "viewer",
                        "collaborator",
                        "owner"
                      ]
                    },
                    "pinned": {
                      "type": "boolean"
                    },
                    "archived": {
                      "type": "boolean"
                    },
                    "isTemplate": {
                      "type": "boolean"
                    },
                    "color": {
                      "type": "string",
                      "enum": [
                        "default",
                        "coral",
                        "peach",
                        "sand",
                        "mint",
                        "sage",
                        "fog",
                        "storm",
                        "dusk",
                        "blossom",
                        "clay",
                        "chalk"
                      ]
                    },
                    "background": {
                      "type": "string",
                      "enum": [
                        "none",
                        "groceries",
                        "food",
                        "music",
                        "recipes",
                        "notes",
                        "places",
                        "travel",
                        "video",
                        "celebration"
                      ]
                    },
                    "position": {
                      "type": "string"
                    },
                    "trashedAt": {
                      "nullable": true,
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                    }
                  },
                  "required": [
                    "id",
                    "type",
                    "title",
                    "bodyHtml",
                    "hasLinks",
                    "items",
                    "labelIds",
                    "attachments",
                    "reminder",
                    "collaborators",
                    "role",
                    "pinned",
                    "archived",
                    "isTemplate",
                    "color",
                    "background",
                    "position",
                    "trashedAt",
                    "createdAt",
                    "updatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/notes/delete-all": {
      "post": {
        "tags": [
          "notes"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "confirm": {
                    "type": "string",
                    "enum": [
                      "delete-all-notes"
                    ]
                  }
                },
                "required": [
                  "confirm"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deleted": {
                      "type": "number"
                    },
                    "left": {
                      "type": "number"
                    },
                    "labels": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "deleted",
                    "left",
                    "labels"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/notes/{id}/versions": {
      "get": {
        "tags": [
          "versions"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                      },
                      "createdBy": {
                        "nullable": true,
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "createdAt",
                      "createdBy"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}